mutual-fund-analyzer/requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0

# Data analysis
scipy>=1.11.0
scikit-learn>=1.3.0

# Web scraping
selenium>=4.15.0
lxml>=4.9.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
openpyxl>=3.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

